% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{check_columns_existence}
\alias{check_columns_existence}
\title{Check whether specified columns exist in a data frame}
\usage{
check_columns_existence(dat, columns)
}
\arguments{
\item{dat}{data.frame to check for the existence of columns}

\item{columns}{a vector of column names (should be "character")}
}
\value{
Throws an error if some of the columns are not present in the
data.frame. The error text indicates which columns are missing
}
\description{
Check whether specified columns exist in a data frame
}
