% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{check_for_delims}
\alias{check_for_delims}
\title{Helper function throwing an error if a string contains characters used
as delimiters in a schema.}
\usage{
check_for_delims(x, delim, x.description, delim.description)
}
\arguments{
\item{x}{character. String to be checked for delimiters}

\item{delim}{character. Delimiter character}

\item{x.description}{Description of the string being checked. Will be
substituted instead of X to the error message: "The following X
contains the character used as Y (see next param)"}

\item{delim.description}{Descriprion of delimiter being checked. Will be
substituted instead of Y in the above error message.}
}
\value{
Nothing
}
\description{
Helper function throwing an error if a string contains characters used
as delimiters in a schema.
}
