% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{check_rois_mois_existence}
\alias{check_rois_mois_existence}
\title{Check whether eye-tracking data contains regions of interest (rois) and
measures of interest (mois).}
\usage{
check_rois_mois_existence(dat, rois, mois, region.col, measure.col)
}
\arguments{
\item{dat}{data.frame to check the existence of rois and mois}

\item{rois, mois}{vectors of region and measures identifiers, which will be
checked against \code{region.col} and \code{measure.col} respectively.
The type of entries in the vector should be the same as in the corresponding
columns}

\item{region.col, measure.col}{name of column containing region
and eye-tracking measures identifiers respectively. See "Details" for more
info.}
}
\value{
Throws an error if some of the rois and mois are not present in the
data.frame; the error text indicates which rois/mois are missing.
}
\description{
Check whether eye-tracking data contains regions of interest (rois) and
measures of interest (mois).
}
\details{
The type of \code{region.col} and \code{measure.col} should be either
"character" or "quosure" (see documentation for "quosure" in \code{rlang}
package)
}
