% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{check_schema_for_delims}
\alias{check_schema_for_delims}
\title{A validation function for a schema checking that names and values of list elements
do not contain characters which will be used as delimiters in the file name}
\usage{
check_schema_for_delims(schema, tag.delim, component.delim)
}
\arguments{
\item{schema}{list. List with a schema to be checked}

\item{tag.delim}{character. The character which will be used as a delimiter
between component tags and values}

\item{component.delim}{character. The character which will be used as a delimiter
between components tags}
}
\value{
Nothing. Throws an error if the naming schema tags or values contain
delimiter characters
}
\description{
A validation function for a schema checking that names and values of list elements
do not contain characters which will be used as delimiters in the file name
}
