% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{count_cells}
\alias{count_cells}
\title{Count number of observations of specific kind}
\usage{
count_cells(dat, by, ..., cast.formula, value.name = "count")
}
\arguments{
\item{dat}{data.frame containing data to process}

\item{by}{unquoted column name for the main grouping variable. See "Details"}

\item{...}{unqouted column names - other grouping variables. See "Details".}

\item{cast.formula}{character. A formula for \code{dcast()} (from 'reshape2' package).
LHS is the grouping variable, RHS defines cells}

\item{value.name}{character. name of the column containing counts in the
result data.frame}
}
\value{
data.frame with 3 columns (default names):
\enumerate{
\item Name of the \code{by} argument, typically subj or item; contains
the unique values from the corresponding column in \code{dat}
\item \code{cell}; contains data subset identifiers. In a typical use,
if used with regions and measures, it would be region_measure, e.g. "critical_tt"
\item \code{count} count of observations in each subset of the data.
}
}
\description{
A generic function to count number of certain observations (e.g. NAs or
extreme values) by cells (e.g. combination of region and eye-tracking measure).
}
\details{
The function assumes that \code{dat} argument contains the data which has
already been subsetted to only contain values of interest. E.g., if you
want to count NAs, before passing the data set to this function, you need
to filter all non-NAs out. The package contains two specific convenience
functions which would do the subseetting for you (the names are self-explanatory):
\code{count_NAs} and \code{count_extremes}.

The \code{by} argument would typically contain subject or item column name.
this is the main grouping variable, which will be displayed on the y axis
in the summary plots.

The column names in \code{...} argument define which columns will act as
grouping variables for \code{dplyr::group_by}, thus defining the smallest
subset of the data in which the observations should be counted. In a typical
use for eye-tracking data, there will be two components here: region and
measure columns names. In general, the values in the first item will be
varied slowest, the values in the last item will be varied fastest.

Thus, e.g. (if \code{by = subj} and \code{... = region.col, measure,col}),
the observations will be counted in each region, in each measure,
for each subject.
}
\seealso{
\code{\link{count_NAs}}, \code{\link{count_extremes}}
}
