% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{count_extremes}
\alias{count_extremes}
\title{Convenience wrapper for count_cells() for counting NAs.}
\usage{
count_extremes(dat, by, rois, mois, value.col, region.col, measure.col,
  max.cutoff, min.cutoff = 0)
}
\arguments{
\item{dat}{data.frame containing data to process}

\item{by}{quosure with column name for the main grouping variable. See "Details"}

\item{value.col}{quosure with column name containing the NAs (potentially along
with non-NA values. Typically it will be the column containing the reaction
times or something alike).}

\item{region.col}{quosure with column name containing region identifiers}

\item{measure.col}{quosure with column name containing measure identifiers}

\item{max.cutoff}{numeric value, indicating the upper threshold; any values
above this threshold will be considered "extreme" and counted}

\item{min.cutoff}{numeric value, indicating the lower threshold. Defaults to 0}
}
\value{
data.frame with 4 columns. The first 3 are analogous to
\code{\link{count_cells()}}
(i.e. they contain the main grouping variable identifier (typically subject or item),
the combination of region and measure identifying subsets of data,
and extreme values counts), and the fourth one, called "direction", indicates
whether the value is extremely high (then the corresponding row will contain
"above.max" in the"direction" column) or extremely low ("below.min" in the
"direction" column).
}
\description{
Count the number of NAs in the data briken down by region, measure and
some third factor, usually subject or item.
}
\details{
The \code{by} argument would typically contain subject or item column name.
this is the main grouping variable, which varies slowest and will be
displayed on the y axis in the summary plots.
}
