% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{count_NAs}
\alias{count_NAs}
\title{Convenience wrapper for count_cells() for counting NAs.}
\usage{
count_NAs(dat, by, rois, mois, value.col, region.col, measure.col)
}
\arguments{
\item{dat}{data.frame containing data to process}

\item{by}{quosure. column name for the main grouping variable. See "Details"}

\item{value.col}{quosure. column name containing the NAs (potentially along
with non-NA values. Typically it will be the column containing the reaction
times or something alike).}

\item{region.col}{quosure. column name containing region identifiers}

\item{measure.col}{quosure. column name containing measure identifiers}
}
\value{
data.frame with 3 columns analogous to \code{\link{count_cells}}
(i.e. containing the main grouping variable identifier (typically subject or item),
the combination of region and measure identifying subsets of data, and NAs counts)
}
\description{
Count the number of NAs in the data briken down by region, measure and
some third factor, usually subject or item.
}
\details{
The \code{by} argument would typically contain subject or item column name.
this is the main grouping variable, which varies slowest and will be
displayed on the y axis in the summary plots.
}
