% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_quest_acc}
\alias{get_quest_acc}
\title{Function to summarize question response accuracy}
\usage{
get_quest_acc(dat, subj.col = "subj", correct.col = "is.correct")
}
\arguments{
\item{dat}{data containing subject identifiers and information about
question answering accuracy. The data has to contain just one region
and one eye-tracking measure in case of eye-tracking data.}

\item{subj.col}{unquoted name of the column containing subject identifiers}

\item{correct.col}{unquoted name of the column containing indicator of whether
the question was answered correctly}
}
\value{
a dataframe with 4 columns: subject number, total number of questions
(will be different if you removed some trials), number of correctly answered
questions, proportion of correct answers.
}
\description{
Function to summarize question response accuracy
}
