% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_subj_info}
\alias{get_subj_info}
\title{A function to extract basic info about subjects from an eye-tracking data
returned by Ilia's scripts.}
\usage{
get_subj_info(dat)
}
\arguments{
\item{dat}{data frame containing eytracking data}
}
\value{
list with two components: \code{n.subj} - number of subjects in the data;
\code{subj.missing} - a vector with IDs of missing subjects; NA if no subjects
are missing. The missing subjects are simply numbers missing from a continuous
vector of numbers from 1 to highest subject ID. E.g. if we have subjects
1,2,3,5, the subject 4 will be reported missing.
@export
}
\description{
A function to extract basic info about subjects from an eye-tracking data
returned by Ilia's scripts.
}
