% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_today}
\alias{get_today}
\title{A function which returns today's date in the requested format. Default format: 02Apr16}
\usage{
get_today(format = "\%d\%b\%y")
}
\arguments{
\item{string.}{Format to use on today's date}
}
\value{
character with today's date
}
\description{
A function which returns today's date in the requested format. Default format: 02Apr16
}
