% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{inflate_region_names}
\alias{inflate_region_names}
\title{Make sure that all regions get a name}
\usage{
inflate_region_names(reg.names, regions, missing.prefix = "Reg_")
}
\arguments{
\item{reg.names}{named character vector, where names correspond to the new
region names to be assigned, and the values correspond to current
region identifiers}

\item{regions}{vector of unique region identifiers in the data}

\item{missing.prefix}{what to prefix to region identifiers for which the user didn't
specify names explicitly. Defaults to "Reg_", i.e. if region identifiers
are numbers, it would generate something like "Reg_1".}
}
\value{
named character vector analogous to \code{reg.names}, but with added
names for regions for which the user didn't specify names
}
\description{
The function ensures that if the user didn't specify some region names
explicitly, some kind of label would still be generated.
}
