% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{make_filename}
\alias{make_filename}
\title{A function to create a filename based on tagging and value schemas.}
\usage{
make_filename(proj.name, tags, values, tag.delim = ".",
  component.delim = "_", data.type)
}
\arguments{
\item{proj.name}{character. Project name; the filename will start with it}

\item{tags}{list. List with a "tagging schema": list elements contain tags for
specific components of the name. E.g. list(markup = "mk",
analysis = "an").}

\item{values}{list. List with a "values schema": lsit elements contain
values for the components of the name associated with a specific
analysis. E.g. list(markup = "like-parker", analysis = "variant4")}

\item{tag.delim}{character. The character which will be used as a delimiter
between component tags and values}

\item{component.delim}{character. The character which will be used as a delimiter
between components tags}

\item{data.type}{optional string specifying type of the data stored in the file.
it will be added after the project name but before the tags. E.g., if
\code{data.type = "data"}, and the tags and values are as in the
examples above, the name could be project1_data_mk.parker-like_an.variant4}
}
\value{
string. Filename, starting with project name
E.g. project1_mk.parker-like_an.variant4
}
\description{
A function to create a filename based on tagging and value schemas.
}
