% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{name_regions_in_subset}
\alias{name_regions_in_subset}
\title{Assign meaningful region names to a subset of data with a single stimuli type}
\usage{
name_regions_in_subset(dat, reg.names, region.col, missing.prefix = "Reg_")
}
\arguments{
\item{dat}{data.frame to add region names to}

\item{reg.names}{named character vector, where names correspond to the new
region names to be assigned, and the values correspond to current
region identifiers}

\item{region.col}{unquoted name of the column containing current region
identifiers}

\item{missing.prefix}{what to prefix to region identifiers for which the user didn't
specify names explicitly. Defaults to "Reg_", i.e. if region identifiers
are numbers, it would generate something like "Reg_1".}
}
\value{
data.frame with the added \code{reg.name} column
}
\description{
A helper function to assign region names to a single type of stimuli.
If data only has one stimuli, this function is called directly to assign
region names; otherwise, it is called repeatedly for each subset of data
corresponding to a stimuli type.
}
