% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{normalize_axes_labels}
\alias{normalize_axes_labels}
\title{A function to make nice axes lables for ggplot.}
\usage{
normalize_axes_labels(s, revert = FALSE)
}
\arguments{
\item{s}{vector of labels to be normalized}

\item{revert}{if FALSE, the symbols _,.,( are replaced with new lines. If TRUE,
the operation is reverted; see "Details"}
}
\value{
vector of normalized labels if \code{revert = FALSE}; vector of labels
without newlines if \code{revert = TRUE}
}
\description{
The function will insert new lines symbol \code{(\\n)} replacing "_", ".",
and before "(", so that long labels can be displayed on multiple lines
}
\details{
Literally undoing the normalization is obviously not possible, since there is
no way of knowing which symbols were replaced with nelines. Thus, the newlines
not adjacent to "(" are replaced with "_"; the newlines preceding "(" are just
deleted.
}
\examples{
labels <- c("critical_region", "spillover.region", "region(another)")
normalize_axes_labels(labels)
# [1] "critical\\nregion"  "spillover\\nregion" "region\\n(another)"
}
