% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{plot_cells_count}
\alias{plot_cells_count}
\title{Plot heatmaps of cell counts}
\usage{
plot_cells_count(dat, by, cells.count, rois, mois, fill = count,
  plot.name = "Give a sensible name to the plot!")
}
\arguments{
\item{dat}{data.frame containing information about \code{by} column}

\item{by}{quosure indicating which column we are aggregating along. Usually
it is the subject or the item column.}

\item{cells.count}{data.frame with cells count data produced by \code{count_cells}.
The function expects that the data.frame has column "cell", containing
the names of data subsets for which counts were performed (e.g
region+measure combination) and column "count", containing the counts
of the relevant values in each cell.}

\item{rois, mois}{vectors of identifiers for regions and measures, in which
cells have to be counted}

\item{fill}{unquoted name of the column with cell counts}

\item{plot.name}{character}
}
\value{
heatmap plot of the cell counts (expanded compared to the plot
returned by \code{plot_cells_heatmap} - with added zero counts and ROIs
delimitation)
}
\description{
A wrapper around \code{\link{plot_cells_heatmap()}} which is aware of regions
and measures of interest. Because of that it a) adds 0 counts to region + measure
combinations which lacked the relevant kind of value (e.g. extreme or NAs);
b) visually delimits ROIs in the plot for visual convenience.
}
