% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{preserve_global_env}
\alias{preserve_global_env}
\title{Copy all objects in the global environment}
\usage{
preserve_global_env()
}
\value{
copy of the global environment
}
\description{
Convenience wrapper around \code{\link{preserve_env}}
}
