% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{report_extremes_count}
\alias{report_extremes_count}
\title{Count and plot extreme observation by regions and measures of interest}
\usage{
report_extremes_count(dat, by = subj, rois, mois, value.col = value,
  region.col = reg.name, measure.col = fixationtype, max.cutoff,
  min.cutoff = 0, plot.name = "Give a sensible name to the plot!")
}
\arguments{
\item{dat}{data.frame containing information about \code{by} column}

\item{by}{quosure indicating which column we are aggregating along. Usually
it is the subject or the item column.}

\item{rois, mois}{vectors of identifiers for regions and measures, in which
cells have to be counted}

\item{value.col}{unquoted name of the column containing NAs which should
be counted (typically it would be the column containing reaction times
or something similar)}

\item{region.col}{unquoted name of the column containing region ids}

\item{measure.col}{unquoted name of the column containing measure ids}

\item{max.cutoff}{numeric value, indicating the upper threshold; any values
above this threshold will be considered "extreme" and counted}

\item{min.cutoff}{numeric value, indicating the lower threshold. Defaults to 0}

\item{fill}{unquoted name of the column with cell counts}
}
\value{
list with 3 components:
\itemize{
\item \code{counts} - data.frame with counts data
\item \code{count.plots} - list with 3 components
\itemize{
\item \code{high.plot} - plot with counts of extreme observations above \code{max.cutoff}
\item \code{low.plot} - plot with counts of extreme observations below \code{min.cutoff}
\item \code{all.plot} - plot with combined counts of both high and low extreme values
}
\item \code{values.plots} - list with 2 components
\itemize{
\item \code{high.plot} - plot with numerical values of extreme observations above \code{max.cutoff}
\item \code{low.plot} - plot with numerical values of extreme observations below \code{min.cutoff}
}
}
}
\description{
Count and plot extreme observation by regions and measures of interest
}
