% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{report_NAs_count}
\alias{report_NAs_count}
\title{Count and plot NAs by regions and measures of interest}
\usage{
report_NAs_count(dat, by = subj, rois, mois, value.col = value,
  region.col = reg.name, measure.col = fixationtype,
  plot.name = "Give a sensible name to the plot!")
}
\arguments{
\item{dat}{data.frame containing the data where the NAs should be counted}

\item{by}{unquoted column name for the main grouping variable. See "Details"}

\item{rois, mois}{vectors of region and measures identifiers, for which counts
should be produced. The type of entries in the vector should be the same as in
\code{region.col} and \code{measure.col} respectively. Any of these can be left
unsepcified, in this case, all regions and/or measure form the data will be used.}

\item{value.col}{unquoted name of the column containing NAs which should
be counted (typically it would be the column containing reaction times
or something similar)}

\item{region.col}{unquoted name of the column containing region ids}

\item{measure.col}{unquoted name of the column containing measure ids}

\item{plot.name}{character. name for the plot visualizing the counts}
}
\value{
list with the following components:
\itemize{
\item counts data.frame containing count data
\item cells.plot heatmap visualization of the counts
}
}
\description{
Count and plot NAs by regions and measures of interest
}
\details{
The \code{by} argument would typically contain subject or item column name.
this is the main grouping variable, which varies slowest and will be
displayed on the y axis in the summary plots.
}
