% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{restore_env}
\alias{restore_env}
\title{Assign all objects from one environment to another.}
\usage{
restore_env(source.env, target.env)
}
\arguments{
\item{source.env}{source environment}

\item{target.env}{target environment}
}
\value{
Nothing, function is called for its side effects
}
\description{
WARNING: objects in the target environments which have the same names as objects
in the source environments will be overwritten!
}
