% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{restore_global_env}
\alias{restore_global_env}
\title{Restore contents of global environment from a saved state}
\usage{
restore_global_env(source.env)
}
\arguments{
\item{source.env}{preserved copy of global environment from which it is to be restored}
}
\value{
Nothing, function is called for its side effects
}
\description{
Convenience wrapper around \code{\link{restore_env}}.
\strong{WARNING}: objects in the target environments which have the same names as objects
in the source environments will be overwritten!
}
