% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{validate_proj_name}
\alias{validate_proj_name}
\title{A validation function for project name checking that it
does not contain characters which will be used as delimiters
in the file name}
\usage{
validate_proj_name(proj.name, tag.delim, component.delim)
}
\arguments{
\item{proj.name}{project name to be validated}

\item{tag.delim}{character. The character which will be used as a delimiter
between component tags and values}

\item{component.delim}{character. The character which will be used as a delimiter
between components tags}
}
\value{
Nothing. Throws an error if the naming project name contains
delimiter characters
}
\description{
A validation function for project name checking that it
does not contain characters which will be used as delimiters
in the file name
}
