% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{validate_schema}
\alias{validate_schema}
\title{General validation function for a naming schema. Just a wrapper,
gathering other validation function in a single place}
\usage{
validate_schema(schema, tag.delim, component.delim,
  check.values.for.dups = FALSE)
}
\arguments{
\item{schema}{list. List with a schema to be checked}

\item{tag.delim}{character. The character which will be used as a delimiter
between component tags and values}

\item{component.delim}{character. The character which will be used as a delimiter
between components tags}

\item{check.values.for.dups}{logical. If TRUE, only names are checked
(usuful for checking values schemas - values can be duplicater);
if FALSE, names and values of the list elements are checked
(useful for checking tagging schemas - tags should be unique)}
}
\value{
Nothing. Throws an error if the naming schema has some problems
}
\description{
General validation function for a naming schema. Just a wrapper,
gathering other validation function in a single place
}
