% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{zero_inflate_cells_count}
\alias{zero_inflate_cells_count}
\title{Make sure that every roi + moi combination has counts}
\usage{
zero_inflate_cells_count(dat, by, cells.count, rois, mois)
}
\arguments{
\item{dat}{data.frame containing information about \code{by} column}

\item{by}{quosure indicating which column we are aggregating along. Usually
it is the subject or the item column.}

\item{cells.count}{data.frame with cells count data produced by \code{count_cells}.}

\item{rois, mois}{vectors of identifiers for regions and measures, in which
cells have to be counted}
}
\value{
data.frame containing the original cells counts as well as added
0 counts
}
\description{
This function ensures that every combination of region and measure
contains counts. Basically, it takes care fo the situation when a given
region-measure doesn't contain cells of the required kind (e.g. NAs or
extreme observations). In such cases this function adds 0 as count to make
sure that all region-measure combination are present in the summary plot.
}
