% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naming_functions.R
\name{check_for_duplicates}
\alias{check_for_duplicates}
\title{A validation helper which throws an error if a vector contains duplicates}
\usage{
check_for_duplicates(x)
}
\arguments{
\item{x}{vector to be checked for duplicates}
}
\description{
A validation helper which throws an error if a vector contains duplicates
}
