% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_today}
\alias{get_today}
\title{A function which returns today's date in the requested format.}
\usage{
get_today(format = "\%d\%b\%y")
}
\arguments{
\item{format}{Format to use on today's date (see \code{\link[base]{strptime}}).
Default format: 02Apr16}
}
\value{
today's date as character
}
\description{
A function which returns today's date in the requested format.
}
