% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{preserve_env}
\alias{preserve_env}
\title{Copy objects from one environment to another, creating a copy of an environment}
\usage{
preserve_env(env = globalenv())
}
\arguments{
\item{env}{source environment (by default - global environment)}
}
\value{
copy of the source environment
}
\description{
Copy objects from one environment to another, creating a copy of an environment
}
